Updated test cases

+++ Models/DTOs/PurchaseResponse.cs
@@ -7,6 +7,6 @@ namespace purchase_service.Models.DTOs;
 public class PurchaseResponse
 {
     [JsonPropertyName("purchase_id")]
-    public int PurchaseId { get; set; }
+    public long PurchaseId { get; set; }
 }
